+++ noodles-bcf/examples/bcf_read_header.rs
@@ -2,14 +2,14 @@
 //!
 //! The result matches the output of `bcftools head <src>`.
 
-use std::{env, fs::File, io};
+use std::{env, io};
 
 use noodles_bcf as bcf;
 
 fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = File::open(src).map(bcf::Reader::new)?;
+    let mut reader = bcf::reader::Builder.build_from_path(src)?;
     let header = reader.read_header()?;
     print!("{header}");