+++ noodles-bcf/examples/bcf_view.rs
@@ -4,7 +4,6 @@
 
 use std::{
     env,
-    fs::File,
     io::{self, BufWriter},
 };
 
@@ -14,7 +13,7 @@ use noodles_vcf as vcf;
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = File::open(src).map(bcf::Reader::new)?;
+    let mut reader = bcf::reader::Builder.build_from_path(src)?;
     let header = reader.read_header()?;
 
     let stdout = io::stdout().lock();